pyproject.toml
[project]
name = "fact-checker"
version = "0.1.0"
description = "The Fact Checker evaluates LLM-generated answers, verifies actual accuracy using the web, and refines the response to ensure alignment with real-world knowledge."

readme = "README.md"

requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "google-adk[eval]>=1.0.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
deployment = [
    "absl-py>=2.2.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
